Core/Scripts: remove forgotten #define in boss_kalecgos.cpp
This #define was originally supposed to be removed in my PR #17832,
but for some reason it remained undeleted. This commit corrects it.

Also aligned all the enum values to the same column for readability,
based on a common multiple of 4 spaces.

(cherry picked from commit 068656af064db741adf92281fd5a094db66b7d4c)